Fix FileComparison dialog
The delete buttons will now cancel the thumbnail tasks, in case they are
still loading the thumbnails and thus using the files. And in case
deleting the file causes and exception, the dialog will not be closed.
To allow the user to handle the situation. Set dialog to fullscreen and
optimized button locations.