Improve performance of loading duplicates
The caching around the images of a VideoFile has been greatly
improved. When adding a file to the list of duplicates, it will try to
copy already loaded images for thumbnail view. Though the images are
now cached as MemoryStreams for WCF serialization. This causes copying
MemoryStream content to Image objects twice on every image compare.