Fix VideoFile serialization
Additionally opening the folder when resolving a duplicate in a client
has been implemented.
When serializing a VideoFile the WCF serializer didn't load the
properties since it directly accessed the member field. Thus the
properties had to be preloaded manually. Additionally multiple issues
with caching the data in the VideoFiles have been fixed, where the
properties didn't properly access their respective field.