Fix not getting enough images
When the video is too short, the stepping, start and end values as
double would cause mpv to round them to the nearest integer. Would could
cause not getting enough images from a single video because the end is
reached too early. Now we round towards zero. But we also keep a min
value of 1 to make sure videos with less seconds than the count of
images don't go into infinite loop.