Add precise seek mode as fallback
It tries to seek images with KeyFramesOnly first. Broken files tend to
miss key frames. Thus, if the MpvWrapper it can't get the number of
images from the key frames, it will change the seek mode to precise.
Which should get the correct number of images if possible. If that
doesn't work, we're all doomed anyways.