Add third level of image loading
When DedupEngine::VideoFile is loading the images, it tries to minimize
the image loads even further. Using three levels of load. First the
minimum necessary for different videos. Then the minimum necessary for
duplicate videos. At last the rare chance were some of the images are
different and some are considered identical.