Fix loading of engine state
When the DedupEngine initialized with settings that do not have any
ExcludedDirectories, it tried to get calculate the hash value of a list
that is null at that point. This is now fixed by adding a null check.

(cherry picked from commit 5e7d013e11ba5d641cf06b42c839e4416f2e1799)