Fix issue when loading images from broken files
When a file is broken we try to load images anyways. If we got enough
images, we stop and return the images. This causes a dispose which
cleans up the process. Which consists of 2 parts. First terminating mpv
and the second is deleting the temporary folder. The order was the wrong
way around. It tried to delete the folder first and terminate mpv
afterwards. This caused an IO Exception because the folder contained
files that are being written to at that moment.