Fix exception in x86 reading from mpv
When reading the height of a video resolution using x86 compiled
version, it would cause an exception because of corrupt memory. This was
due to the fact that an IntPtr was used as the type for the long. While
mpv was promised it would be 64bit. While in fact it was only 32bit when
compiled as x86. Now it's using the proper type long (aka int64) for
this parameter. Though we fixed reading of the string to clean up the
code.