Improve image comparison performance
Instead of parallelising the image comparison, we now prepare the images
properly. Since we are reusing the same images many times. This way, the
comparison itself doesn't have to do it. And since the comparison
actually uses a scaled down to 16x16 pixel image in greyscale, we store
a lot of memory too. Since we don't even store the original image
anymore. The XnaFan ImageComparison lib has been replaced. Since we
basically rewrote a big part of what we used from it. But the algorithm
is still the same.