Fix duration label growing to the right
When the duration label displayed a duration larger than a day, it would
grow to the right, thus into the right hand margin of the dialog. This
is fixed by using a TableLayoutPanel with RightToLeft active, AuotSize
active, Anchor set to top-right and Margins reduced to 0. This makes the
label and the TableLayoutPanel grow to the left properly.