Add image bytes to stored engine state
We do store the image bytes of all images that have been loaded for all
the video in the engine state. Thus speeding up the comparison when
continuing. Since the engine state file is now larger, we don't store it
after every compared file anymore. Instead it's stored regularly in a
specifiable interval.