Improve compatibility of engine state
When trying to load engine state depending on the current configuration,
we try to consider less settings. Meaning, it doesn't matter when the
MaxImageDifferencePercent has changed. Since we re-compare all images on
every start now. The only settings that are considered still are the
basePath and the MaxImageComapares. Because those require extracting the
images from the videos again. Which is the process that actually takes
time. Since we re-compare all images on every start, we don't store the
current index of comparison anymore and we don't store fund duplicates.
Since we find them again in a few seconds.