Add cropping of black bars on images
Since some videos might have mattes (black bars) at the top and bottom
or left and right side, it causes issues when comparing the images with
those of another video. Since part of the image is just black and highly
increases the similarities. Even though the black bars don't add
information we care about. Thus we remove those black bars when
pre-processing the images for comparison.