Fix handling of broken video files
When trying to load images of broken video files, it would get stuck
because the mpv handle of the file was closed already. In addition, it
returns null for images that are unable to be exported. Thus the
comparison and display for those images has been fixed with proper
null-checks.