Fix handling of display of broken file comparison
For the ImageComparisonResultView as well as FilePreviewDlg, when one of
the images couldn't be loaded, we now show a specific image to indicate
issues when loading the file. And the text in the
ImageComparisonResultView is improved to show more clearly that one or
both of the files couldn't be loaded and the comparison was skipped.