Add double buffering for VideoDedupDlg
Activated double buffering for the main form of the VideoDedupClient.
This is supposed to help with flickering of the log when getting a lot
of entries. Though I didn't see any issues without double buffering
recently. So I couldn't confirm. But it shouldn't be of an issue to
activate it.