Optimized image extraction with libmpv
Instead of using a start, stop and interval to let libmpv iterate over
the file itself, we now pause the video and then jump to the specified
time stamps. This also allows us to extract non-consecutive images from
a video. And it's also faster and more reliable and doesn't require
multiple tries with different seek modes.