Add database store to replace engine state
Instead of having the engine state stored in json files, we use an
SQLite DB now. Additionally we changed the order of the images we
extract from files. Instead of starting from the front, we use the
fractions in which we divide the videos and sort those by their
fractions. This way the chance is higher to reuse images when the image
count to compare has changed.