Performance improve for reading duration
When the DedupEngine::VideoFile had issues reading the duration using
MpvWrapper, it would retry every time the Duration property was used to
get the value. But since it's very unlikely that it would work the
second time around AND because it would cause inconsistent behavior, we
stay with the value we got first. In the case of a broken file, it's
TimeSpan.Zero.