Improve performance of imageIndices cache
Since we stored the imageIndices as IOrderedEnumerable, the Enumerable
actually had to be resolved every time to check the count of images.
Now, instead using the List, the List itself takes care of caching the
count. And the Enumerable will be resolved as soon as we write the
imageIndices cache.