Replace ImageGroupBox with custom controls
The client area of the ImageGroupBox didn't calculate properly when
AutoSize was true. The higher the header the more of the lower part of
the client area was cut off since the calculation for the AutoSize
property didn't account for the higher header. I couldn't find a proper
fix so I replaced the whole control with a TableLayoutPanel that has a
PictureBox and Label in the top column with the GroupBox of the load
level in the lower column. With the GroupBox actually not having its
text property set and acting merely as a visual framing, it still looks
almost the same but works better.