Improve performance of VideoComparisonPreviewDlg
Improved performance by suspending the layout of TableLayoutPanels as
well as moving the conversion of the MemoryStream containing the images
to Image and resizing them into a background thread. Both drastically
reduces the work required to be done in the UI thread. Further
improvements could be done by consolidating the nested TableLayoutPanels
and replacing the Winforms.Timer with a Threading.Timer.