Add additional status information on operations
Instead of the main UI Dialog of VideoDedupClient, the display and
handling of operation info is no moved to a separate custom control.
This way, the custom control can handle the specifics about how the
information is formatted and displayed. The assignment of status info
text to the OperationType values is now handles in that custom control
as well. Thus, the server will only send OperationTypes instead of
display text.