Add File.Exists() check in MpvWrapper
Duplicates for deleted files could be found, due to cached files in the
Datastore. But when loading the preview images for the duplicate, it
would try to access an empty file. The idea was to check the duration
for Zero. But since the duration was cached as well, the MpvWrapper
tried to read from the non-existing file. Which actually triggered
multiple events from the libmpv. Like TrackChanged, StartFile, EndFile
and Idle. But it ended in a timeout. This delayed the processing
greatly. Now added File.Exists(...) checks for all public files that
access the file with libmpv. This throws the newly created
MpvFileNotFoundException.