Fix sorting of ImageIndex
Since it's a grpc protoc class, protoc will already provide the hash and
comparison functions for the class. So we removed our custom
implementation, which was a left over from the WCF implementation.
Instead, we use OrderBy and ThenBy at the place where we need the custom
ordering. That is enough for now.