Add ignore of Warning ICE61
It complains about being able to uninstall new versions with an older
version of a setup. Which is true. Because we want to be able to be able
to uninstall a 1.0.8124.31244 with a 1.0.8124.31245 version. And that
seems to only be possible with the AllowSameVersionUpgrades. Because it
just checks the first two parts of the version. It's especially
confusing because the name of the AllowSameVersionUpgrades flag.