Fix reset of position in MemoryStream
When we create the CacheableImageSet, we create the MemoryStream of the
bytes to create an image from it. We reset the position after creating
the image. But when the image is used to create the cropped, small and
greyscaled versions, it moves the position of the MemoryStream again. So
we have to make sure to set the Position after any operations on the
image have been completed.