Fix display of difference type in ServerConfigDlg
When the dialog was opened and the duration difference type Percent was
selected, the event handler to display the proper unit for the
difference type wasn't called. Because it was only called when the
RdbDurationDifferenceSeconds was changed. But since it defaults to false
and was changed to true, it didn't trigger the event handler. Now the
event handler is triggered for both radio boxes.