Fix deleting wrong file of duplicate
The VideoComparisonDlg would switch left and right video if the file
size of the right video was higher than the file size of the left video.
Making sure the left video was either the same size of larger than the
right one. But when selecting to delete either the left or right file,
it would send the ResolveOperation to delete either the left or right
file to the server. But since the server didn't know about the dialog
switching the files, it would delete the wrong file. This is solved by
the ResolveDuplicateRequest having an explicit reference to the file to
delete.