Fix setup by updating libs and GUIDs
Added missing libs for the replacement of System.Data.Sqlite with
Microsoft.Data.Sqlite and the new MpvLib and VideoComparer. And updated
the component GUIDs because the installer wasn't able to uninstall the
service anymore. It's not 100% clear why that happened. But the GUIDs
have probably been copied from another project and having both installed
screwed with registry entries.