Specify backoff behavior for grpc client
The client uses a specific backoff configuration for all messages except
the GetCurrentStatus message. The default is 5 retries with an initial
delay of up to 1 second, a multiplier of 1.2 and a maximum backoff of 5
seconds for StatusCode Unavailable. For the GetCurrentStatus message,
it's 2 retries with an initial delay of up to 0.1 seconds with a
multiplier of 1 and a maximum backoff of 0.1 seconds for StatusCode
Unavailable.