Add PathBox with text and button view
The PathBox shows the path, similar to the explorer path box as either
an array of buttons as well as a TextBox if needed. The array of buttons
is dynamic to the width of the PathBox. So if the size of the PathBox
shrinks it will hide the left most buttons until at least one button is
visible or the buttons are all displayed.