Fix ResizableButtonArray buttons
The folder buttons on the path bar didn't work when the user switched to
a completely different path, because the buttons still had their
original callback lambdas with references to the original elements.
Those are now properly loaded from the tags of the buttons. And the
quickSelectButton now gets it's tag properly updated. That wasn't event
he case before.