Fix address bar when reusing dialog
If the CustomSelectFileDialog is reused, the address bar could reorder
the buttons incorrectly. Reason is the buttons are reused and get new
text values when the path changes. But the order of the buttons in
TlpArray.Controls changes and does not reflect the order in which the
buttons are displayed.