Add underscore to event handlers
Added underscores between the controls name and the events name. The
handler HandleDgvContentCellDoubleClick was the culprit. I interpreted
it as handling the ContentCellDoubleClick event. Which is not the case.
It's the CellDoubleClick event of the DgvContent control. The underscore
makes this clear. There goes my conviction of hating on underscores in
PascalCase.