Fix duplicate speed
Since the user can solve duplicates while search is running, we reduce
the current DuplicateCound. This screwed with the duplicate speed
statistic. So we added a new DuplicatesFound to the StatusData. The
server continuously increases the DuplicatesFound count, until a new
search is started. This way, the count never decreases and is
independent from the DuplicateCount which is now called
CurrentDuplicateCount to make the distinction easier.