Fix MaximumFiles of OperationInfo
We can't always use e.MaxCount to assign MaximumFiles. When the
operation changes to monitoring or waiting, we need to have the correct
maximum files of the previous operation. So in those cases, we take the
FileCount of the last ProgressInfo. This might cause issues when the
operation was canceled though.