Fix adding empty ProgressInfos
When the deduplication operation is finished, we get an OperationUpdate
from the DedupEngine that states the new OperationType. Though this also
caused a new ProgressInfo to be added to progressInfos. The client would
then try to display it, causing issues, since it's not part of the
operation. whose ProgressInfos we're trying to display.