Fix PrepareCandidates
The function was still implemented as if it was iterating over a single
list. Which isn't the case anymore. So we can't just skip all videos
that have a lower duration than the video we're looking for candidates
for. Instead we have to consider all known videos as potential
candidates, except itself. It also means, we can't easily stop when the
IsDurationEqual function returns false. Since it happens for shorter
videos already. So we might just compare the duration with the first
video, since it's too short, we would just stop looking for any further
candidates. And since we have to iterate over the whole list of known
video no matter what, we don't need to order them. Nothing gained by
that anymore.