Upgraded Setup to WiX Toolset 5
Biggest change is the use of the Files element, which automatically
harvests files from their respective sources. Combined with the
publishing of VideoDedupClient and VideoDedupService, which is triggered
in the pre-build event of Setup, the files are easily harvested together
for the installer. WiX Toolset 6 is bugged. But the code should work
fine with 6 when it is fixed. Problem is the Files elements. Those cause
issues with 6 when multiple project references exist.