Fix crash when selecting source path
If the source path is empty, CustomSelectFileDialogs internal
currentPath will be null. This bubbles out of CustomSelectFileDialog
through ContentRequested with args.Path that is null. Not supplying any
content, since the path is null, will cause a subsequent exception being
thrown in CustomSelectFileDialog. For now, we just set the CurrentPath
to the first RootEntry if the args.Path is null.