Add SystemInfo
The server now provides a grpc request to get the system information.
This allows us to show that on the client side in the server
configuration. As well as having the tool tip for the ConcurrencyLevel
of the DedupSettingsCtrl show the proper information of the processor
count of the server.