Fix crash when changing server address
When the user changes the server address while the status thread is
requesting the current status, the current status was forwarded to the
StiProgress to display the status. But since the this makes another
request, it will try and request from another server or fail if a
connection isn't possible. Both causing issues, up to a crash of the
client. This is fixed by making sure the server address is still the
same after the current status request.